requirements.txt
awsebcli==3.7.8
boto==2.42.0
botocore==1.4.52
cement==2.8.2
colorama==0.3.7
coverage==4.2
diff-match-patch==20121119
Django==1.10.1
django-import-export==0.5.0
django-mathfilters==0.4.0
django-storages==1.5.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.4
mysqlclient==1.3.7
python-dateutil==2.5.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tablib==0.11.2
texttable==0.8.4
websocket-client==0.37.0
